[network-builder] move setup_network out of libra-node/main_node.rs
