+++ network/builder/Cargo.toml
@@ -15,6 +15,7 @@ bytes = { version = "0.5.4", features = ["serde"] }
 futures = "0.3.5"
 once_cell = "1.4.0"
 pin-project = "0.4.17"
+proptest = { version = "0.10.0", default-features = true, optional = true }
 rand = "0.7.3"
 serde = { version = "1.0.111", default-features = false }
 serde_bytes = "0.11.4"
@@ -34,6 +35,7 @@ libra-metrics = { path = "../../common/metrics", version = "0.1.0" }
 libra-network-address = { path = "../../network/network-address", version = "0.1.0" }
 libra-proptest-helpers = { path = "../../common/proptest-helpers", version = "0.1.0", optional = true }
 libra-security-logger = { path = "../../common/security-logger", version = "0.1.0" }
+libra-secure-storage = {path = "../../secure/storage", version = "0.1.0"}
 libra-types = { path = "../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 memsocket = { path = "../../network/memsocket", version = "0.1.0" }
@@ -41,6 +43,6 @@ netcore = { path = "../../network/netcore", version = "0.1.0" }
 network = {path = "../../network", version = "0.1.0"}
 num-variants = { path = "../../common/num-variants", version = "0.1.0" }
 onchain-discovery = {path = "../../network/onchain-discovery", version = "0.1.0"}
-proptest = { version = "0.10.0", default-features = true, optional = true }
+storage-interface = {path = "../../storage/storage-interface", version = "0.1.0"}
 stream-ratelimiter = { path = "../../common/stream-ratelimiter", version = "0.1.0" }
 rand_core = { version = "0.5.1", optional = true }
