+++ network/builder/src/network_builder.rs
@@ -12,14 +12,15 @@
 use channel::{self, libra_channel, message_queues::QueueStyle};
 use futures::stream::StreamExt;
 use libra_config::{
-    config::{RoleType, HANDSHAKE_VERSION},
+    config::{DiscoveryMethod, NetworkConfig, RoleType, HANDSHAKE_VERSION},
     network_id::NetworkId,
 };
 use libra_crypto::x25519;
 use libra_logger::prelude::*;
 use libra_metrics::IntCounterVec;
 use libra_network_address::NetworkAddress;
-use libra_types::PeerId;
+use libra_secure_storage::config;
+use libra_types::{waypoint::Waypoint, PeerId};
 use netcore::transport::Transport;
 use network::{
     common::NetworkPublicKeys,
@@ -46,7 +47,11 @@ use std::{
     sync::{Arc, RwLock},
     time::Duration,
 };
-use tokio::{runtime::Handle, time::interval};
+use storage_interface::DbReader;
+use tokio::{
+    runtime::{Builder, Handle, Runtime},
+    time::interval,
+};
 use tokio_retry::strategy::ExponentialBackoff;
 
 // TODO:  AuthenticationModes should be defined lower in the technical stack;  the mode is useful
@@ -517,4 +522,99 @@ impl NetworkBuilder {
             EventT::new(peer_mgr_reqs_rx, connection_notifs_rx),
         )
     }
+}
+
+// TODO(abhayb): Move to network crate (similar to consensus).
+pub fn setup_network(
+    config: &mut NetworkConfig,
+    role: RoleType,
+    libra_db: Arc<dyn DbReader>,
+    waypoint: Waypoint,
+) -> (Runtime, NetworkBuilder) {
+    let runtime = Builder::new()
+        .thread_name("network-")
+        .threaded_scheduler()
+        .enable_all()
+        .build()
+        .expect("Failed to start runtime. Won't be able to start networking.");
+
+    let identity_key = config::identity_key(config);
+    let peer_id = config::peer_id(config);
+
+    let mut network_builder = NetworkBuilder::new(
+        runtime.handle().clone(),
+        config.network_id.clone(),
+        peer_id,
+        role,
+        config.listen_address.clone(),
+    );
+    network_builder.add_connection_monitoring();
+
+    if config.enable_remote_authentication {
+        // Sanity check seed peer addresses.
+        config
+            .seed_peers
+            .verify_libranet_addrs()
+            .expect("Seed peer addresses must be well-formed");
+
+        let network_peers = config.network_peers.peers.clone();
+        let seed_peers = config.seed_peers.seed_peers.clone();
+
+        let trusted_peers = if role == RoleType::Validator {
+            // for validators, trusted_peers is empty will be populated from consensus
+            HashMap::new()
+        } else {
+            network_peers
+        };
+
+        info!(
+            "network setup: role: {}, seed_peers: {:?}, trusted_peers: {:?}",
+            role, seed_peers, trusted_peers,
+        );
+
+        network_builder
+            .advertised_address(config.advertised_address.clone())
+            .authentication_mode(AuthenticationMode::Mutual(identity_key))
+            .trusted_peers(trusted_peers)
+            .seed_peers(seed_peers)
+            .connectivity_check_interval_ms(config.connectivity_check_interval_ms)
+            // TODO:  Why is the connectivity manager related to remote_authentication?
+            .add_connectivity_manager();
+    } else {
+        // Even if a network end-point operates without remote authentication, it might want to prove
+        // its identity to another peer it connects to. For this, we use TCP + Noise but without
+        // enforcing a trusted peers set.
+        network_builder
+            .authentication_mode(AuthenticationMode::ServerOnly(identity_key))
+            .advertised_address(config.advertised_address.clone());
+    }
+
+    match config.discovery_method {
+        DiscoveryMethod::Gossip => {
+            network_builder
+                .discovery_interval_ms(config.discovery_interval_ms)
+                .add_gossip_discovery();
+        }
+        DiscoveryMethod::Onchain => {
+            let (network_tx, discovery_events) =
+                network_builder.add_protocol_handler(onchain_discovery::endpoint_config());
+            let discovery_builder =
+                onchain_discovery::discovery_builder::OnchainDiscoveryBuilder::build(
+                    network_builder
+                        .conn_mgr_reqs_tx()
+                        .expect("ConnecitivtyManager not enabled"),
+                    network_tx,
+                    discovery_events,
+                    config::peer_id(config),
+                    role,
+                    libra_db,
+                    waypoint,
+                    runtime.handle(),
+                );
+            discovery_builder.start(runtime.handle());
+        }
+        DiscoveryMethod::None => {}
+    }
+
+    (runtime, network_builder)
 }
